+++ src/promise/HISTORY.md
@@ -11,7 +11,12 @@ Promise Change History
 * The first function received as parameter of the Promise init function is now
   a reference to resolve() instead of fulfill()
 * Changed the value of |this| inside callbacks to undefined to match the
-Promises A+ spec
+  Promises A+ spec
+
+3.10.1
+------
+
+* No changes.
 
 3.10.0
 ------